% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anitools.R
\name{ANI.histogram}
\alias{ANI.histogram}
\title{Construct a histogram using ANI values}
\usage{
ANI.histogram(
  bactaxRObject,
  bindwidth = 0.01,
  xmin = 79,
  xmax = 100,
  xline = NULL,
  xlinecol = "#20A387FF",
  xlinetype = "dashed"
)
}
\arguments{
\item{bactaxRObject}{A bactaxRObject}

\item{xmin}{Minimum value for the X-axis. Defaults to 79 (ANI).}

\item{xmax}{Maximum value for the X-axis. Defaults to 100 (ANI).}

\item{xline}{Numeric value (or vector of numeric values) at which
to draw vertical line(s) along the X-axis (i.e., ANI values).
Defaults to NULL (no vertical lines).}

\item{xlinecol}{Character (or vector of characters) corresponding to
ggplot2 color(s) for xline. Defaults to "#20A387FF".}

\item{xlinetype}{Character (or vector of characters) corresponding to
ggplot2 line type(s) for xline. Defaults to "dashed".}

\item{binwidth}{Width of bins in the histogram. Defaults to 0.01 (ANI).}
}
\value{
A ggplot2 object
}
\description{
This function constructs a histogram using ANI values
}
