% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anitools.R
\name{load.ANI}
\alias{load.ANI}
\title{Load dataframe or matrix, and store it as a bactaxRObject object}
\usage{
load.ANI(x)
}
\arguments{
\item{x}{Matrix or data frame, where Columns 1, 2, and 3 contain the query genome,
reference genome, and associated ANI values, respectively (remaining columns will be ignored).}
}
\value{
A bactaxRObject
}
\description{
This function takes a matrix or data frame already stored in memory,
and converts it to a bactaxR object. The matrix or data frame should contain at
least 3 columns, where Column 1 contains a query genome, Column 2 contains a
reference genome, and Column 3 contains ANI values (ranging from 0 to 100)
This matrix or data frame must contain pairwise comparisons for all query and reference genomes
 (i.e., all-vs-all comparisons).
}
