% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anitools.R
\name{read.ANI}
\alias{read.ANI}
\title{Read pairwise output from fastANI}
\usage{
read.ANI(file)
}
\arguments{
\item{file}{Path to file containing pairwise ANI comparisons output by fastANI
 (or any 3+ column file, described as follows).
The file should not have a header, and should contain 3 or more columns,
each separated by a tab. Columns 1, 2, and 3 should contain the query genome,
reference genome, and the associated ANI value (ranging from 0 to 100), respectively
(remaining columns will be ignored).}
}
\value{
A bactaxRObject object
}
\description{
This function reads a file of pairwise ANI values calculated by fastANI
(or a file of pairwise ANI values constructed using other methods; see below).
This file must contain pairwise comparisons for all query and reference genomes
 (i.e., all-vs-all comparisons).
}
